moved expression templates to a separate header
